refactor: relocate components to `ui` for better structure
- Moved all components to the `ui` folder to improve file structure.
- Updated imports across the project to reflect the new paths.
- Adjusted mock paths in test files to align with the updated structure.
