+++ src/components/Hero/Hero.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import Hero from "@/components/Hero/Hero";
 
-jest.mock('../SociaisLinks/SocialLinks', () => {
+jest.mock('@/components/ui/SociaisLinks/SocialLinks', () => {
   const MockSocialLinks = () => <div data-testid="social-links"/>;
   MockSocialLinks.displayName = 'MockSocialLinks';
   return MockSocialLinks;
@@ -10,7 +10,7 @@ jest.mock('../SociaisLinks/SocialLinks', () => {
 describe('Hero Component', () => {
   it('should render a section with the ID "home"', () => {
     render(<Hero/>);
-    const section = screen.getByTestId('hero');
+    const section = screen.getByTestId('home');
     expect(section).toBeInTheDocument();
     expect(section).toHaveAttribute('id', 'home');
   });
