+++ src/components/ui/HeroImage/HeroImage.tsx
